% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_val.R
\name{alpha_val}
\alias{alpha_val}
\title{bernouli distribution over edge existance probability}
\usage{
alpha_val(rho)
}
\arguments{
\item{rho}{Prior over edge existance}
}
\value{

}
\description{
bernouli distribution over edge existance probability
}
