% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_states.R
\name{compute_states}
\alias{compute_states}
\title{call to ODE solver}
\usage{
compute_states(W, q, r0, times, method, ts, flag, stimulation.value)
}
\arguments{
\item{W}{wighted matrix}

\item{q}{perturbation vector (1 = perturbed, 0 = unperturbed)}

\item{r0}{initial state of unperturbed system}

\item{times}{number of time points}

\item{method}{which case (SteadyState or TimeSeries)}

\item{ts}{time step length}

\item{flag}{flag to plot states}

\item{stimulation.value}{constant value for stimulation function}
}
\value{
matrix of states for given experiment
}
\description{
call to ODE solver
}
