% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulation.R
\name{data_simulation}
\alias{data_simulation}
\title{simulate data for a model}
\usage{
data_simulation(
  model,
  nE,
  times,
  theta,
  ts,
  method,
  method.wt,
  depletion.factor.wt,
  stimulation.value,
  stimulation.value.wt
)
}
\arguments{
\item{model}{list :W, Q, prior.theta, prior.W, orig.dat, r0}

\item{nE}{number of observations to simulate}

\item{times}{number of times points to simulate}

\item{theta}{S_E connection prior}

\item{ts}{time steps length}

\item{method}{time series or steady state}

\item{depletion.factor.wt}{diag of W in WT}

\item{stimulation.value}{constant value for stimulation function}

\item{stimulation.value.wt}{constant value for stimulation function in WildType case}
}
\value{
data array (nE x #number of experiments x times)
}
\description{
simulate data for a model
}
