% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_effect.R
\name{density_effect}
\alias{density_effect}
\title{density of effect expected}
\usage{
density_effect(d, gumb_loc, gumb_scl, gumb_shp)
}
\arguments{
\item{d}{a numeric quantile}

\item{gumb_loc}{location parameter of extereme distribution}

\item{gumb_scl}{scale parameter of extereme distribution}

\item{gumb_shp}{shape parameter of extereme distribution}
}
\value{

}
\description{
density of effect expected
}
