% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_no_effect.R
\name{density_no_effect}
\alias{density_no_effect}
\title{density of no effect expected}
\usage{
density_no_effect(d, trunc_mean, trunc_sd)
}
\arguments{
\item{d}{a numeric quantile}

\item{trunc_mean}{mean of truncated normal}

\item{trunc_sd}{standard deviation of truncated normal}
}
\value{

}
\description{
density of no effect expected
}
