% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammas.R
\name{gammas}
\alias{gammas}
\title{mixture coefficient of measurement distribution}
\usage{
gammas(R, R_0, flag)
}
\arguments{
\item{R}{states under experiment}

\item{R_0}{states in WildType}

\item{flag}{flag for printing plots}
}
\value{
level of activity
}
\description{
mixture coefficient of measurement distribution
}
