% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_layer_ODEsolver.R
\name{hidden_layer_ODEsolver}
\alias{hidden_layer_ODEsolver}
\title{main ODE solver}
\usage{
hidden_layer_ODEsolver(t, x, parms, input)
}
\arguments{
\item{t}{time}

\item{x}{ODE answer}

\item{parms}{list of input parameters: experiment vector,W}

\item{input}{stimulation function}
}
\value{
dynamic of hidden nodes under special experiment
}
\description{
main ODE solver
}
