% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_prior.R
\name{log_prior}
\alias{log_prior}
\title{compute log-prior}
\usage{
log_prior(W, prior.W)
}
\arguments{
\item{W}{wights of hidden structure}

\item{prior.W}{list of rho, lambda, nu, tau = hyperparameters of prior over W}
}
\value{

}
\description{
compute log-prior
}
