% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_sampling.R
\name{MCMC_sampling}
\alias{MCMC_sampling}
\title{learn W from data}
\usage{
MCMC_sampling(
  dat,
  Q,
  r_0,
  rho,
  hyper.W,
  prior.theta,
  data.para,
  depletion.factor,
  depletion.factor.wt,
  method.wt,
  stimulation.value,
  stimulation.value.wt,
  seed = seed,
  ts,
  itr
)
}
\arguments{
\item{dat}{feeded data of correspondings experiments,dimension nObs.nExp.times}

\item{Q}{matrix of perturbation vectors Experiments*hidden nodes}

\item{r_0}{start states of hidden nodes}

\item{rho}{prior over edge existance}

\item{hyper.W}{list of rho, lambda, nu, tau = hyperparameters of prior over W}

\item{prior.theta}{prior probabilities over attachements E-nodes to S-nodes}

\item{depletion.factor}{diag of W}

\item{depletion.factor.wt}{diag of W in WT}

\item{method.wt}{method for WildType state calculation: steadyState or timeSeries}

\item{stimulation.value}{constant value for stimulation function}

\item{stimulation.value.wt}{constant value for stimulation function in WildType case}

\item{seed}{random seed}

\item{ts}{Timesteps}

\item{itr}{number of iterations for MCMC}
}
\value{
returns sampled Ws and LogLiklihoods
}
\description{
learn W from data
}
