% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_generator.R
\name{model_generator}
\alias{model_generator}
\title{constructing Model Object}
\usage{
model_generator(
  stype,
  ti_num,
  prior.theta,
  prior.W,
  data.para,
  W.factor,
  depletion.factor,
  depletion.factor.wt,
  init.ODE,
  E_num,
  method.wt,
  stimulation.value,
  stimulation.value.wt,
  tstep
)
}
\arguments{
\item{stype}{name of motif : "Diamond","Bifan","FeedForward","FeedBackward","ProteinCascade"}

\item{ti_num}{number of time points}

\item{prior.theta}{prior over observations and hidden nodes connections}

\item{prior.W}{prior over W}

\item{data.para}{list of parameter for data simulation: mu, sd, location, scale shape}

\item{W.factor}{W=1 * W.factor}

\item{depletion.factor}{diag of W}

\item{depletion.factor.wt}{diag of W in WT}

\item{init.ODE}{initial value of ODE}

\item{E_num}{number of observations}

\item{method.wt}{method for WildType state calculation: steadyState or timeSeries}

\item{stimulation.value}{constant value for stimulation function}

\item{stimulation.value.wt}{constant value for stimulation function in WildType case}

\item{tstep}{time step length}
}
\value{

}
\description{
constructing Model Object
}
