% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_generator_effect.R
\name{random_generator_effect}
\alias{random_generator_effect}
\title{random generator of observation when an effect is expected}
\usage{
random_generator_effect(gumb_loc, gumb_scl, gumb_shp)
}
\arguments{
\item{gumb_loc}{location parameter of extereme distribution}

\item{gumb_scl}{scale parameter of extereme distribution}

\item{gumb_shp}{shape parameter of extereme distribution}
}
\value{
one randome value far from zero
}
\description{
random generator of observation when an effect is expected
}
