% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_generator_no_effect.R
\name{random_generator_no_effect}
\alias{random_generator_no_effect}
\title{random generator of observation when no effect expected}
\usage{
random_generator_no_effect(trunc_mean, trunc_sd)
}
\arguments{
\item{trunc_mean}{mean of truncated normal}

\item{trunc_sd}{standard deviation of truncated normal}
}
\value{

}
\description{
random generator of observation when no effect expected
}
