% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_network.R
\name{sample_network}
\alias{sample_network}
\title{construct a network structure}
\usage{
sample_network(S_num, P_num, E_num, theta, g_s)
}
\arguments{
\item{S_num}{number of Hidden nodes}

\item{P_num}{number of perturbation nodes}

\item{E_num}{number of Observations nodes per each hidden nodes}

\item{theta}{matrix of prior information for S_E connections}

\item{g_s}{hidden layer network structure}
}
\value{
an igraph of all nodes and connection as true network
}
\description{
construct a network structure
}
