% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimulation.R
\name{stimulation}
\alias{stimulation}
\title{Stimulation function}
\usage{
stimulation(stim_val, t, len)
}
\arguments{
\item{stim_val}{constant value for stimulation function}

\item{t}{specific time point}

\item{len}{number of hidden nodes}
}
\value{
decreasing or uprising vector during time
}
\description{
Stimulation function
}
