% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_matrix_generator.R
\name{weight_matrix_generator}
\alias{weight_matrix_generator}
\title{W generator}
\usage{
weight_matrix_generator(alpha, prior.W, small_val)
}
\arguments{
\item{alpha}{confidence of edge existance}

\item{prior.W}{list of rho, lambda, nu, tau = hyperparameters of prior over W}

\item{small_vall}{diag value}
}
\value{

}
\description{
W generator
}
