% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildtype_states.R
\name{wildtype_states}
\alias{wildtype_states}
\title{WildeType State}
\usage{
wildtype_states(model, r0, times, ts, flag)
}
\arguments{
\item{model}{list :W, Q, prior.theta, prior.W, orig.dat, r0}

\item{times}{number of times points to simulate}

\item{ts}{time steps length}

\item{flag}{flag to plot states}
}
\value{

}
\description{
WildeType State
}
